# docs/contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Agent Vue FastAPI Contract
  version: 1.0.0
  description: |
    本文件已随设计基线 65c08f32f5f61b74791b260a02fd62ce3fe91bd1 获得用户批准，
    是 `/api/v1` 唯一机器可读契约；实现、前端类型、契约测试和文档必须与批准版本一致。
    项目采用单租户简化身份，业务角色为 `user`、`reviewer` 与 `admin`；对象级授权仍由服务端执行。
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      summary: 注册用户并签发访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionEnvelope"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/login:
    post:
      tags: [Authentication]
      operationId: loginUser
      summary: 登录并签发访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/sms-codes:
    post:
      tags: [Authentication]
      operationId: sendSmsCode
      summary: 发送注册或找回密码验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SmsCodeRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SmsCodeEnvelope" }
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
  /auth/password-resets:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: 使用短信验证码重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordResetEnvelope" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
...
  /auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      summary: 获取当前登录用户资料
      x-required-role: user
      responses:
        "200":
          description: 当前用户资料
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Authentication]
      operationId: updateCurrentUser
      summary: 更新当前登录用户的昵称或头像
      x-required-role: user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: 用户资料已更新
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/me/password:
    patch:
      tags: [Authentication]
      operationId: changeCurrentUserPassword
      summary: 修改当前用户密码
      x-required-role: user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: 密码已更新
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/me/avatar:
    put:
      tags: [Authentication]
      operationId: uploadCurrentUserAvatar
      summary: 上传当前用户头像
      x-required-role: user
      requestBody:
        required: true
        content:
          image/jpeg:
            schema: { type: string, format: binary }
          image/png:
            schema: { type: string, format: binary }
          image/webp:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: 头像已更新
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /knowledge/files:
    post:
      tags: [Knowledge]
      operationId: uploadKnowledgeFile
      summary: 上传知识库文本文件
      x-required-role: admin
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 180
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string, format: binary }
          text/markdown:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: 知识文件已上传
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
    get:
      tags: [Knowledge]
      operationId: listKnowledgeFiles
      summary: 列出已入库的知识文件
      x-required-role: admin
      responses:
        "200":
          description: 知识文件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeFilePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /knowledge/reindex:
    post:
      tags: [Knowledge]
      operationId: reindexKnowledgeFiles
      summary: 重建已上传知识文件的检索索引
      x-required-role: admin
      responses:
        "200":
          description: 已完成索引重建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeReindexEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
...
  /me/usage-summary:
    get:
      tags: [Usage Summaries]
      operationId: getCurrentUserUsageSummary
      summary: 获取当前用户的平台使用总结
      description: |
        返回由真实站内对话、长期记忆和已推荐型号汇总的最新快照。
        服务端从访问令牌确定用户身份，不接受 user_id 参数；设备业务数据缺失时返回明确状态，不会编造。
      x-required-role: user
      responses:
        "200":
          description: 当前用户总结快照；数据不足时返回 status=empty。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummaryEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
  /me/usage-events:
    post:
      tags: [Usage Summaries]
      operationId: recordCurrentUserUsageEvent
      summary: 记录当前用户的产品浏览行为
      description: 仅允许记录详情页和 3D 查看行为；用户身份从访问令牌确定，写入数据严格按 user_id 隔离。
      x-required-role: user
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageEventRequest"
      responses:
        "201":
          description: 行为已记录
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageEventEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
...
  /sessions/{session_id}:
    get:
      tags: [Sessions]
      operationId: getSession
      summary: 获取当前用户的会话详情
      x-required-role: user
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: 会话详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/StorageUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        可替换认证 Adapter 验证令牌并提供 `sub` 与 `role`；role 只能是 `user`、`reviewer` 或 `admin`。
        reviewer 具备普通 user 能力，但不得绕过对象级授权读取无关用户资源。
...
  responses:
...
    ValidationError:
      description: 请求结构校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [phone, password, nickname, verification_code, user_agreement_version, privacy_policy_version, agree_user_agreement, agree_privacy_policy]
      properties:
        phone:
          type: string
          minLength: 11
          maxLength: 16
        password:
          type: string
          minLength: 6
          maxLength: 20
          pattern: '^(?=.*[A-Za-z])(?=.*[0-9])(?=.*[^A-Za-z0-9]).{6,20}$'
        nickname:
          type: string
          minLength: 1
          maxLength: 40
        verification_code:
          type: string
          pattern: '^[0-9]{6}$'
        user_agreement_version:
          type: string
          minLength: 1
          maxLength: 32
        privacy_policy_version:
          type: string
          minLength: 1
          maxLength: 32
        agree_user_agreement:
          type: boolean
          const: true
        agree_privacy_policy:
          type: boolean
          const: true
        avatar_url:
          type: [string, "null"]
          minLength: 8
          maxLength: 2048
          default: null
...